BRIGHT estimation procedure for individual level data

Fit BRIGHT estimation procedure for individual level data with grouped penalties over a grid of values of the regularization parameter lambda.

BRIGHTi(
  X_plink,
  Y,
  group = 1:ncol(X),
  Beta_prior,
  penalty = "grLasso",
  family = "gaussian",
  eta,
  lambda,
  lambda.min,
  tau = 1/3,
  alpha = 1,
  gamma = ifelse(penalty == "grSCAD", 4, 3),
  eps = 1e-07,
  max_iter = 1e+06,
  dfmax = 5000,
  gmax = 5000
)

Arguments

group

A vector describing the grouping of the coefficients. For greatest efficiency and least ambiguity (see details), it is best if group is a factor or vector of consecutive integers, although unordered groups and character vectors are also allowed. If there are coefficients to be included in the model without being penalized, assign them to group 0 (or "0").

Beta_prior

a numeric verticle vector of the prior information; in the case where prior information is a subspace of XtY, set the complement set to be zero.

penalty

the penalty to be applied to the model. For group selection, one of grLasso, grMCP, or grSCAD.

family

either "gaussian" or "binomial", depending on the response.

eta

a numeric scalar of the weights for incorporating prior information.

lambda

a user supplied sequence of lambda values. Typically, this is left unspecified, and the function automatically computes a grid of lambda values that ranges uniformly on the log scale over the relevant range of lambda values.

lambda.min

the smallest value for lambda, as a fraction of lambda.max. Default is .0001 if the number of observations is larger than the number of covariates and .05 otherwise.

tau

tuning parameter for the group exponential lasso; defaults to 1/3.

alpha

grpreg allows for both a group penalty and an L2 (ridge) penalty; alpha controls the proportional weight of the regularization parameters of these two penalties. The group penalties' regularization parameter is lambda*alpha, while the regularization parameter of the ridge penalty is lambda*(1-alpha). Default is 1: no ridge penalty.

gamma

tuning parameter of the group or composite MCP/SCAD penalty (see details). Default is 3 for MCP and 4 for SCAD.

eps

convergence threshhold. The algorithm iterates until the RMSD for the change in linear predictors for each coefficient is less than eps. Default is 1e-4. See details.

dfmax

limit on the number of parameters allowed to be nonzero. If this limit is exceeded, the algorithm will exit early from the regularization path.

gmax

limit on the number of groups allowed to have nonzero elements. If this limit is exceeded, the algorithm will exit early from the regularization path.

X_bed

the genotype matrix directory stored as plink .bed format. BRIGHT automatically load it and standardizes the data.

y

the response vector.

max.iter

maximum number of iterations (total across entire path). Default is 10000. See details.

Value

An object with S3 class "grpreg" containing:

beta

The fitted matrix of coefficients. The number of rows is equal to the number of coefficients, and the number of columns is equal to nlambda.

family

Same as above.

group

Same as above.

lambda

The sequence of lambda values in the path.

alpha

Same as above.

loss

A vector containing either the residual sum of squares (`"gaussian"`) or negative log-likelihood (`"binomial"`) of the fitted model at each value of `lambda`.

n

Number of observations.

penalty

Same as above.

df

A vector of length `nlambda` containing estimates of effective number of model parameters all the points along the regularization path. For details on how this is calculated, see Breheny and Huang (2009).

iter

A vector of length `nlambda` containing the number of iterations until convergence at each value of `lambda`.

group.multiplier

A named vector containing the multiplicative constant applied to each group's penalty.

Details

For more information about the penalties and their properties, please consult the references below, many of which contain discussion, case studies, and simulation studies comparing the methods. If you use BRIGHT for an analysis, please cite the appropriate reference.

In keeping with the notation from the original MCP paper, the tuning parameter of the MCP penalty is denoted 'gamma'. Note, however, that in Breheny and Huang (2009), gamma is denoted 'a'.

The objective function for BRIGHTi optimization is defined to be $$Q(\beta)=||y-X\beta||/2n+||X\beta'-X\beta||/2n+p(\beta),$$ where the first term on the right hand side (RHS) is the OLS loss; the second term on the RHS is the Bregman-divergence; the third term on the RHS is the penalty.

The algorithms employed by BRIGHT are stable and generally converge quite rapidly to values close to the solution. However, especially when p is large compared with n, BRIGHT may fail to converge at low values of lambda, where models are nonidentifiable or nearly singular. Often, this is not the region of the coefficient path that is most interesting. The default behavior warning the user when convergence criteria are not met may be distracting in these cases, and can be modified with warn (convergence can always be checked later by inspecting the value of iter).

If models are not converging, increasing max.iter may not be the most efficient way to correct this problem. Consider increasing n.lambda or lambda.min in addition to increasing max.iter.

Although BRIGHT allows groups to be unordered and given arbitary names, it is recommended that you specify groups as consecutive integers. The first reason is efficiency: if groups are out of order, X must be reordered prior to fitting, then this process reversed to return coefficients according to the original order of X. This is inefficient if X is very large. The second reason is ambiguity with respect to other arguments such as group.multiplier. With consecutive integers, group=3 unambiguously denotes the third element of group.multiplier.

BRIGHT requires groups to be non-overlapping.

References

Li Q, Patrick MT, Zhang H, Khunsriraksakul C, Stuart PE, Gudjonsson JE, Nair R, Elder JT, Liu DJ, Kang J, Tsoi LC. Bregman Divergence-Based Data Integration with Application to Polygenic Risk Score (PRS) Heterogeneity Adjustment. arXiv preprint arXiv:2210.06025. 2022 Oct 12.

Breheny P and Huang J. (2009) Penalized methods for bi-level variable selection. *Statistics and its interface*, **2**: 369-380. doi: 10.4310/sii.2009.v2.n3.a10

Huang J, Breheny P, and Ma S. (2012). A selective review of group selection in high dimensional models. *Statistical Science*, **27**: 481-499. doi: 10.1214/12-sts392

Breheny P and Huang J. (2015) Group descent algorithms for nonconvex penalized linear and logistic regression models with grouped predictors. *Statistics and Computing*, **25**: 173-187. doi: 10.1007/s11222-013-9424-2

Breheny P. (2015) The group exponential lasso for bi-level variable selection. *Biometrics*, **71**: 731-740. doi: 10.1111/biom.12300

See also

Author

Qinmengge Li