BRIGHT estimation procedure for summary level data

Fit BRIGHT estimation procedure for summary level data (BRIGHTs) with grouped penalties (LASSO, elastic net, MCP, SCAD) over a grid of values of the regularization parameter lambda and the prior data weight eta.

BRIGHTs(
  data,
  m = NA,
  group = NA,
  penalty = 1,
  tau = 1/3,
  eta = c(0, exp(seq(log(0.1), log(10), length.out = 20))),
  lambda = NA,
  nlambda = 100,
  lambda.min = 0.001,
  alpha = 1,
  gamma = 0,
  eps = 1e-07,
  max_iter = 1e+06,
  dfmax = 5000,
  gmax = 5000,
  user = T
)

Arguments

data

a list of preprocessed data from PreprocessS(). For details please refer to the PreprocessS() section.

m

a numeric verticle vector of the multiplicative factor to adjust for the number of SNPs in each group, the default is the square root of number of elements in each group.

group

a vector describing the grouping of the coefficients. For greatest efficiency and least ambiguity (see details), it is best if group is a factor or vector of consecutive integers, although unordered groups and character vectors are also allowed. If there are coefficients to be included in the model without being penalized, assign them to group 0 (or "0").

penalty

a integer scalar for chosing the penalties; 1 corresponds to elastic net (with alpha=0 to be LASSO); 2 corresponds to MCP; 3 corresponds to SCAD.

tau

tuning parameter for the group exponential lasso; defaults to 1/3.

eta

a numeric scalar of the weights for incorporating prior information.

lambda

a numeric verticle vector of the LASSO penalty weights; it must be sorted with a decreasing order.

alpha

BRIGHT allows for both a group penalty and an L2 (ridge) penalty; alpha controls the proportional weight of the regularization parameters of these two penalties. The group penalties' regularization parameter is lambda*alpha, while the regularization parameter of the ridge penalty is lambda*(1-alpha). Default is 1: no ridge penalty.

gamma

tuning parameter of the group or composite MCP/SCAD penalty (see details). Default is 3 for MCP and 4 for SCAD.

eps

convergence threshhold. The algorithm iterates until the RMSD for the change in linear predictors for each coefficient is less than eps. Default is 1e-4. See details.

dfmax

limit on the number of parameters allowed to be nonzero. If this limit is exceeded, the algorithm will exit early from the regularization path.

gmax

limit on the number of groups allowed to have nonzero elements. If this limit is exceeded, the algorithm will exit early from the regularization path.

user

a logical scalar for indicating whether lambda is user specified; if user=TRUE, then the iteration will start from the largest lambda, otherwise the iteration will start from the second largest lambda.

max.iter

maximum number of iterations (total across entire path). Default is 10000. See details.

Value

Write a list of results containing: beta, the coefficient estimate from BRIGHT;

iter, the number of total iterations needed for the model to converge with each lambda;

df total degree of freedom of the converged model with each lambda;

dev, the approximated deviance associated with each lambda.

Details

For more information about the penalties and their properties, please consult the references below, many of which contain discussion, case studies, and simulation studies comparing the methods. If you use BRIGHT for an analysis, please cite the appropriate reference.

In keeping with the notation from the original MCP paper, the tuning parameter of the MCP penalty is denoted 'gamma'. Note, however, that in Breheny and Huang (2009), gamma is denoted 'a'.

The objective function for BRIGHTs optimization is defined to be $$Q(\beta)=r\beta+\beta\Sigma\beta/2+(\beta'-\beta)\Sigma(\beta'-\beta)/2+p(\beta),$$ where the first two terms on the right hand side (RHS) are a approximation of the OLS loss; the second term on the RHS is the Bregman-divergence; the third term on the RHS is the penalty.

The algorithms employed by BRIGHT are stable and generally converge quite rapidly to values close to the solution. However, especially when p is large compared with n, BRIGHT may fail to converge at low values of lambda, where models are nonidentifiable or nearly singular. Often, this is not the region of the coefficient path that is most interesting. The default behavior warning the user when convergence criteria are not met may be distracting in these cases, and can be modified with warn (convergence can always be checked later by inspecting the value of iter).

If models are not converging, increasing max.iter may not be the most efficient way to correct this problem. Consider increasing n.lambda or lambda.min in addition to increasing max.iter.

Although BRIGHT allows groups to be unordered and given arbitary names, it is recommended that you specify groups as consecutive integers. The first reason is efficiency: if groups are out of order, X must be reordered prior to fitting, then this process reversed to return coefficients according to the original order of X. This is inefficient if X is very large. The second reason is ambiguity with respect to other arguments such as group.multiplier. With consecutive integers, group=3 unambiguously denotes the third element of group.multiplier.

BRIGHT requires groups to be non-overlapping.