Maximum Lambda selection

Calculate the maximum lambda, which corresponding to penalizing all coefficients to be zero

MaxLambda(XtY, tilde_beta, X, K1, m, blk, K0, tau, eta, alpha, eps, max_iter)

Arguments

XtY

a numeric verticle vector of the marginal correlation between genotype and outcome, can be recovered from GWAS summary statistics by function p2cor().

tilde_beta

a numeric verticle vector of the prior information; in the case where prior information is a subspace of XtY, set the complement set to be zero.

X

a numeric matrix of reference genotypes, default is from 1000 genome project.

K1

a numeric verticle vector of the grouping of SNPs for group penalties; SNPs within K1[i,i+1] are grouped together.

m

a numeric verticle vector of the multiplicative factor to adjust for the number of SNPs in each group, the default is the square root of number of elements in each group.

blk

a numeric verticle vector indicating the grouping of SNPs in each block, can be generated through LD() function.

K0

a integer scalar for the number of SNPs/covariates that is not penalized.

tau

a numeric scalar represents the element-wise soft-thresholding parameter for generating the LD.

eta

a numeric scalar of the weights for incorporating prior information.

alpha

a numeric scalar of the ridge penalty weights; alpha=1 corresponds to no ridge penalty.

eps

a numeric scalar for the convergence criteria.

max_iter

an integer scalar for the maximum iterations before the algorithm stops.

lambda

a numeric verticle vector of the LASSO penalty weights; it must be sorted with a decreasing order.