Preprocess for summary statistics (BRIGHTs)

Preprocess the GWAS and the reference. Quality control is performed first for both target and prior (please see QC section). Then prior "GWAS" or "Correlations" is converted to "Coefficients" through BRIGHTs with eta=0; the hyperparameters are tuned through BRCp.

PreprocessS(
  Tss,
  Tind,
  Tref,
  Pss = NA,
  Pind = NA,
  Pref = NA,
  TLDblocks = "EUR.hg19",
  PLDblocks = NA,
  TN = NA,
  PN = NA
)

Arguments

Tss

a data frame of the target summary statistics, including "GWAS" or "Corr" corresponding to the GWAS or correlation summary statistics, the input summary type must match with Tind below.

Tind

a character scaler of the input target summary type, it must be either "GWAS" or "Corr" corresponding to the GWAS or correlation summary statistics. It must match with the summary statistics in Tss.

Tref

a character scaler indicating the LD reference. It takes values as either a directory to user provided Plink 1 binary files (.bim, .bed, .fam) or the 1000 genome project reference provided by the package: "EUR.hg19", "EUR.hg38", "AFR.hg19", "AFR.hg38", "EAS.hg19", "EAS.hg38", "AMR.hg19", "AMR.hg38", "SAS.hg19" or "SAS.hg38". The abbreviations before dot are populations (more detials can be found here: https://useast.ensembl.org/Help/Faq?id=532); The abbreviations after dot are the genetic build (more details can be found: https://gatk.broadinstitute.org/hc/en-us/articles/360035890951-Human-genome-reference-builds-GRCh38-or-hg38-b37-hg19). For the user provided reference, the SNPs must match with those provided in Pref with the same order.

Pss

a data frame of the prior summary statistics, including "GWAS", "Corr" or "Coef" corresponding to the GWAS, correlation or coefficient summary statistics, the input summary type must match with Pind below.

Pind

a character scaler of the input prior summary type, it must be either "GWAS", "Corr" or "Coef" corresponding to the GWAS, correlation or coefficient summary statistics. It must match with the summary statistics in Pss.

Pref

a character scaler indicating the LD reference. It takes values as either a directory to user provided Plink 1 binary files (.bim, .bed, .fam) or the 1000 genome project reference provided by the package: "EUR.hg19", "EUR.hg38", "AFR.hg19", "AFR.hg38", "EAS.hg19", "EAS.hg38", "AMR.hg19", "AMR.hg38", "SAS.hg19" or "SAS.hg38". The abbreviations before dot are populations (more detials can be found here: https://useast.ensembl.org/Help/Faq?id=532); The abbreviations after dot are the genetic build (more details can be found: https://gatk.broadinstitute.org/hc/en-us/articles/360035890951-Human-genome-reference-builds-GRCh38-or-hg38-b37-hg19). For the user provided reference, the SNPs must match with those provided in Tref with the same order.

TLDblocks

a character scaler indicating the LD block structure to be used from Berisa et al for the target population. It takes values from "EUR.hg19", "EUR.hg38", "AFR.hg19", "AFR.hg38", "EAS.hg19", "EAS.hg38", "AMR.hg19", "AMR.hg38", "SAS.hg19" or "SAS.hg38". The abbreviations before dot are populations (more detials can be found here: https://useast.ensembl.org/Help/Faq?id=532); The abbreviations after dot are the genetic build (more details can be found: https://gatk.broadinstitute.org/hc/en-us/articles/360035890951-Human-genome-reference-builds-GRCh38-or-hg38-b37-hg19). The genetic build must match with those provided in Tref.

PLDblocks

a character scaler indicating the LD block structure to be used from Berisa et al for the prior population. It takes values from "EUR.hg19", "EUR.hg38", "AFR.hg19", "AFR.hg38", "EAS.hg19", "EAS.hg38", "AMR.hg19", "AMR.hg38", "SAS.hg19" or "SAS.hg38". The abbreviations before dot are populations (more detials can be found here: https://useast.ensembl.org/Help/Faq?id=532); The abbreviations after dot are the genetic build (more details can be found: https://gatk.broadinstitute.org/hc/en-us/articles/360035890951-Human-genome-reference-builds-GRCh38-or-hg38-b37-hg19). The genetic build must match with those provided in Pref.

TN

a integer scaler indicating the sample sizes of the target summary statistics input in Tss.

PN

a integer scaler indicating the sample sizes of the target summary statistics input in Tss.

Value

A list of preprocessed data that can be directly input to BRIGHTs().