BRIGHT estimation procedure

Gradient-descent algorithm for optimizing the BRIGHT estimation procedure

gdfit_gaussian(
  XtY,
  tilde_beta,
  X,
  lambda,
  K1,
  m,
  blk,
  K0,
  penalty,
  tau,
  eta,
  alpha,
  gamma,
  eps,
  max_iter,
  dfmax,
  gmax,
  user
)

Arguments

XtY

a numeric verticle vector of the marginal correlation between genotype and outcome, can be recovered from GWAS summary statistics by function p2cor().

tilde_beta

a numeric verticle vector of the prior information; in the case where prior information is a subspace of XtY, set the complement set to be zero.

X

a numeric matrix of reference genotypes, default is from 1000 genome project.

lambda

a numeric verticle vector of the LASSO penalty weights; it must be sorted with a decreasing order.

K1

a numeric verticle vector of the grouping of SNPs for group penalties; SNPs within K1[i,i+1] are grouped together.

m

a numeric verticle vector of the multiplicative factor to adjust for the number of SNPs in each group, the default is the square root of number of elements in each group.

blk

a numeric verticle vector indicating the grouping of SNPs in each block, can be generated through LD() function.

K0

a integer scalar for the number of SNPs/covariates that is not penalized.

penalty

a integer scalar for chosing the penalties; 1 corresponds to LASSO, 2 corresponds to MCP, 3 corresponds to SCAD.

tau

a numeric scalar represents the element-wise soft-thresholding parameter for generating the LD.

eta

a numeric scalar of the weights for incorporating prior information.

alpha

a numeric scalar of the ridge penalty weights; alpha=1 corresponds to no ridge penalty.

gamma

a numeric scalar of the third parameter in firm and SCAD thresholding functions.

eps

a numeric scalar for the convergence criteria.

max_iter

an integer scalar for the maximum iterations before the algorithm stops.

dfmax

an integer scalar for the maximum number of SNPs been selected before the algorithm stops.

gmax

an integer scalar for the maximum number of groups been selected before the algorithm stops.

user

a logical scalar for indicating whether lambda is user specified; if user=TRUE, then the iteration will start from the largest lambda, otherwise the iteration will start from the second largest lambda.

Value

Write a list of results; beta, the coefficient estimate from BRIGHT;

iter, the number of total iterations needed for the model to converge with each lambda;

df total degree of freedom of the converged model with each lambda;

dev, the approximated deviance associated with each lambda.